jobs/requirements.txt
pika
face-recognition
torch
numpy==1.26.4
transformers==4.41.0
python-dotenv